Me falta devolver el token

+++ todoListApi/controllers/todo.cotrollers.js
@@ -1,23 +1,22 @@
 const todoCTRL = {};
 
-const Todo = require('../models/Todo');
+const Todo = require('../model/Todo');
 
 todoCTRL.crearTodo = async (req, res) => {
 
-    const { title, description } = req.body;
-
+    const { title, description, user } = req.body;
 
     try {
         
         const newTodo = new Todo({
-            title, description
+            title, description, user
         });
 
         await newTodo.save();
 
         res.status(201).json({
             ok: true,
-            todo
+            newTodo
         });
 
     } catch (error) {
@@ -31,11 +30,19 @@ todoCTRL.crearTodo = async (req, res) => {
 };
 
 todoCTRL.getTodos = async (req, res) => {
+
+    const {page, limit} = req.query;
+    //Hacemos skip para saltar los primeros elementos si la pagina es mayor a 1
+    const skip = (page - 1) * limit;
+
     try {
-        const todos = await Todo.find();
+        const todos = await Todo.find().skip(skip).limit(limit);
         res.status(200).json({
             ok: true,
-            todos
+            data: todos,
+            page,
+            limit,
+            total: todos.length
         });
     } catch (error) {
         console.log(error)
@@ -50,9 +57,19 @@ todoCTRL.putTodo = async ( req, res ) => {
     const { id } = req.params;
 
     const { title, description } = req.body;
+    const { user } = req.headers;
 
     try {
-        const updateTodo = await Todo.findByIdAndUpdate(id, {title, description}, {new: true});
+        const verificar = await Todo.findOne({_id: id});
+        console.log(verificar)
+        if (verificar.user !== user){
+            return res.status(401).send({
+                ok: false,
+                message: 'No tiene permisos para modificar el Todo'
+            });
+        };
+
+        const updateTodo = await Todo.findByIdAndUpdate({_id: id}, {title, description}, {new: true});
 
         res.status(200).json({
             ok: true,
@@ -64,15 +81,25 @@ todoCTRL.putTodo = async ( req, res ) => {
             ok: false,
             message: 'Error al actualizar el todo'
         });
-    }
+    };
 };
 
 todoCTRL.deleteTodo = async (req, res) => {
     const { id } = req.params;
+    const { user } = req.headers;
 
     try {
 
-        const deleteTodo = await Todo.findByIdAndDelete(id);
+        const verificar = await Todo.findOne({_id: id});
+    
+        if (verificar.user !== user){
+            return res.status(401).send({
+                ok: false,
+                message: 'No tiene permisos para modificar el Todo'
+            });
+        };
+
+        await Todo.findByIdAndDelete(id);
 
         res.status(200).json({
             ok: true,
